extract main to repo root for easy install with
