+++ .gitignore
@@ -1,1 +1,1 @@
-/src/main/main
+golang-mud
